requirements.txt
asgiref==3.4.1
authentication==1.1.0
click==8.0.1
colorama==0.4.4
Django==3.2.7
djangorestframework==3.12.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
passlib==1.7.4
pycryptodome==3.10.1
PyJWT==2.1.0
pytz==2021.1
redis==3.5.3
sqlparse==0.4.2
Werkzeug==2.0.1
WTForms==2.3.3
